Aligns the logic in the welcome resources
as a result the quarkus one can be removed

closes keycloak#23243
