+++ server-spi/src/main/java/org/keycloak/urls/UrlType.java
@@ -2,6 +2,6 @@ package org.keycloak.urls;
 
 public enum UrlType {
 
-    FRONTEND, BACKEND, ADMIN
+    FRONTEND, BACKEND, ADMIN, LOCAL_ADMIN
 
 }
